Changed usage of size_t to uint32_t and ssizt_t to int64_t. Did other miscellaneous changes.

+++ Containers/TC_Allocator.h
@@ -6,11 +6,11 @@ typedef struct TGeneric TGeneric;
 
 typedef struct TC_Allocator
 {
-    void* (*Alloc)(size_t size);
-    void* (*Realloc)(void* memory, size_t new_Size);
-    void* (*Calloc)(size_t num_Elements, size_t size_Of_Elements);
-    void* (*Memcpy)(void* destination, const void * __restrict__ source, size_t size_Of);
-    void* (*Memmove)(void *_Dst, const void *_Src, size_t size_Of);
+    void* (*Alloc)(uint32_t size);
+    void* (*Realloc)(void* memory, uint32_t new_Size);
+    void* (*Calloc)(uint32_t num_Elements, uint32_t size_Of_Elements);
+    void* (*Memcpy)(void* destination, const void * __restrict__ source, uint32_t size_Of);
+    void* (*Memmove)(void *_Dst, const void *_Src, uint32_t size_Of);
     TGeneric (*Free)(TContainer* container);
 } TC_Allocator;
